+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinPlayerCollector.kt
@@ -6,6 +6,7 @@ import com.bitmovin.analytics.BitmovinAnalyticsConfig
 import com.bitmovin.analytics.DefaultCollector
 import com.bitmovin.analytics.adapters.PlayerAdapter
 import com.bitmovin.analytics.bitmovin.player.features.BitmovinFeatureFactory
+import com.bitmovin.analytics.bitmovin.player.providers.PlayerLicenseProvider
 import com.bitmovin.analytics.config.SourceMetadata
 import com.bitmovin.analytics.data.DeviceInformationProvider
 import com.bitmovin.analytics.data.EventDataFactory
@@ -49,6 +50,7 @@ class BitmovinPlayerCollector
         val userAgentProvider = PlayerUserAgentProvider(context, getPlayerAgent())
         val eventDataFactory = EventDataFactory(config, userIdProvider, userAgentProvider)
         val deviceInformationProvider = DeviceInformationProvider(context)
+        val playerLicenseProvider = PlayerLicenseProvider(context)
         return BitmovinSdkAdapter(
             player,
             config,
@@ -57,6 +59,7 @@ class BitmovinPlayerCollector
             sourceMetadataMap,
             eventDataFactory,
             deviceInformationProvider,
+            playerLicenseProvider,
         )
     }