+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinSdkAdapter.kt
@@ -4,6 +4,7 @@ import android.util.Log
 import com.bitmovin.analytics.BitmovinAnalyticsConfig
 import com.bitmovin.analytics.adapters.AdAdapter
 import com.bitmovin.analytics.adapters.DefaultPlayerAdapter
+import com.bitmovin.analytics.bitmovin.player.providers.PlayerLicenseProvider
 import com.bitmovin.analytics.config.SourceMetadata
 import com.bitmovin.analytics.data.DeviceInformationProvider
 import com.bitmovin.analytics.data.ErrorCode
@@ -43,6 +44,7 @@ class BitmovinSdkAdapter(
     private val sourceMetadataMap: Map<Source, SourceMetadata>,
     eventDataFactory: EventDataFactory,
     deviceInformationProvider: DeviceInformationProvider,
+    private val playerLicenseProvider: PlayerLicenseProvider,
 ) : DefaultPlayerAdapter(
     config,
     eventDataFactory,
@@ -229,9 +231,10 @@ class BitmovinSdkAdapter(
             data.audioLanguage = audioTrack.language
         }
 
-        // Player Key
+        // we fall back to using the key from the player config or manifest
+        // in case it is not specified in the analytics config
         if (config.playerKey.isBlank()) {
-            data.playerKey = player.config.key
+            data.playerKey = playerLicenseProvider.getBitmovinPlayerLicenseKey(player.config)
         }
     }